+++ src/sftp.jl
@@ -142,8 +142,8 @@ mutable struct SftpSession
     - [`LibSSHException`](@ref): If creating the SFTP session fails.
     """
     function SftpSession(session::Session)
-        if !isopen(session)
-            throw(ArgumentError("Session is closed, cannot create an SftpSession with it"))
+        if !isconnected(session)
+            throw(ArgumentError("Session is disconnected, cannot create an SftpSession with it"))
         end
 
         ptr = @lockandblock session lib.sftp_new(session.ptr)
@@ -907,6 +907,9 @@ function Base.read(file::SftpFile, nb::Integer=typemax(Int))
     return out
 end
 
+"$(TYPEDSIGNATURES)"
+Base.read(filename::AbstractString, sftp::SftpSession) = open(read, filename, sftp)
+
 """
 $(TYPEDSIGNATURES)
 
@@ -972,6 +975,10 @@ $(TYPEDSIGNATURES)
 Read the whole file as a `String`.
 """
 Base.read(file::SftpFile, ::Type{String}) = String(read(file))
+
+"$(TYPEDSIGNATURES)"
+Base.read(filename::AbstractString, sftp::SftpSession, ::Type{String}) = String(read(filename, sftp))
+
 
 """
 $(TYPEDSIGNATURES)